Update README.md
Kilts are a traditional Scottish dress that has been worn for centuries. They are typically made of tartan fabric and consist of a skirt-like garment that is wrapped around the waist and secured with a belt or buckle. In recent years, kilts have become increasingly popular as a fashion item and can be seen being worn by men around the world. 

Kilts for men is a renowned brand of mens kilt in the USA. We have both traditional and modern day kilts. Modern mens kilts that we have feature unique designs such as utility kilt, tactical kilt and leather kilt and hybrid kilt. Quality of stuff we are using is guaranteed and 100% genuine. 

Whether for a formal event or casual wear, Kilts For Men can provide a stylish and unique addition to any wardrobe. We provide only custom made kilts in the USA. All kilts are handmade, made by our skilled artisans. We pay attention to each customer's order/kilt, each kilt is hand woven and meets the quality standards. You will feel the quality when a kilt will be in your hand. Explore from our wide range of mens kilt.

The Kilt is specifically crafted to fit your waist, hip, and length measurements. For measuring the kilt size properly, please go to our measuring guide page .